match JFWDisplay on J

+++ include/JSystem/JUtility/JUTVideo.h
@@ -33,7 +33,7 @@ public:
         width = (u16)getFbWidth();
         height = (u16)getEfbHeight();
     }
-    u16 getXfbHeight() const { return mRenderObj->xfb_height; }
+    u32 getXfbHeight() const { return mRenderObj->xfb_height; }
     u32 isAntiAliasing() const { return mRenderObj->antialiasing; }
     Pattern getSamplePattern() const { return mRenderObj->sample_pattern; }
     u8* getVFilter() const { return mRenderObj->vfilter; }
